pcsx2: gcc warning round 3 * various cast was overflowing  - use unsigned integer for SSE mask  - force SINGLE macro to use u32 * disable parenthesis warning, only a matter of coding style
Remains 2 wrong use of unsigned integer:
pcsx2/CDVD/InputIsoFile.cpp:96:21: warning: comparison of unsigned expression < 0 is always false [-Wtype-limits]
pcsx2/Memory.cpp:108:86: warning: narrowing conversion of ‘-1’ from ‘int’ to ‘vtlbHandler {aka unsigned int}’ inside { } is ill-formed in C++11 [-Wnarrowing]


git-svn-id: http://pcsx2.googlecode.com/svn/trunk@5689 96395faa-99c1-11dd-bbfe-3dabce05a288

+++ pcsx2/IPU/mpeg2lib/Idct.cpp
@@ -213,7 +213,7 @@ __ri void mpeg2_idct_add (const int last, s16 * block, s16 * dest, const int str
     }
     else
     {
-		int DC = (block[0] + 4) >> 3;
+		s16 DC = (block[0] + 4) >> 3;
 		s16 dcf[2] = { DC, DC };
 		block[0] = block[63] = 0;
